esgbot/config.py
import os
from dotenv import load_dotenv

load_dotenv()

BOT_TOKEN = os.getenv("BOT_TOKEN")
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

if not BOT_TOKEN:
    raise ValueError("BOT_TOKEN не задан в .env файле")
if not OPENAI_API_KEY:
    raise ValueError("OPENAI_API_KEY не задан в .env файле")