esgbot/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
feedparser==6.0.12
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
openai==2.24.0
playwright==1.58.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dotenv==1.2.2
python-telegram-bot==22.6
requests==2.32.5
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.8.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
